requirements.txt
altgraph==0.17.2
Brotli==1.0.9
certifi==2022.6.15
charset-normalizer==2.1.0
docutils==0.19
future==0.18.2
idna==3.3
mutagen==1.45.1
pefile==2022.5.30
Pillow==9.2.0
pycryptodomex==3.15.0
Pygments==2.12.0
pyinstaller==5.2
pyinstaller-hooks-contrib==2022.8
pypiwin32==223
pywin32==304
pywin32-ctypes==0.2.0
requests==2.28.1
urllib3==1.26.11
websockets==10.3
youtube-dl==2021.12.17
yt-dlp==2022.7.18
